+++ src/Uno.UI/UI/Xaml/Controls/Unsupported/RefreshContainer.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Windows.UI.Xaml.Controls;
+
+[Obsolete(
+	"The Windows.UI.Xaml.Controls version of this control is not supported. " +
+	"Please use Microsoft.UI.Xaml.Controls.RefreshContainer instead.")]
+public partial class RefreshContainer
+{
+	public RefreshContainer()
+	{
+		throw new NotImplementedException(
+			"The Windows.UI.Xaml.Controls version of this control is not supported. " +
+			"Please use Microsoft.UI.Xaml.Controls.RefreshContainer instead.");
+	}
+}
